feat(audio): add whisper audio transcription; fix bugs; update docs and CI
Audio transcription:
- Add audio_model_transcribe() accepting TEXT (file path) or BLOB input
- Auto-detect audio format from file extension and magic bytes (WAV, MP3, FLAC)
- Convert audio to mono 16kHz PCM as required by Whisper
- Support 15 transcription options (language, translate, n_threads, temperature, etc.)
- Uncomment and register audio_model_load/audio_model_free functions
- Add 8 new unit tests covering transcription, options, error handling, and lifecycle

Bug fixes:
- Fix strncasecmp prefix-matching bug: add KEY_MATCHES macro that checks both
  key length and case-insensitive match, replacing all 39 occurrences
- Fix wrong variable in logger call: warn_buf was formatted but buffer was passed
  to ai_logger, logging raw option values instead of the warning message
- Fix NULL dereference in role_normalize when role is NULL
- Fix NULL dereference in llm_model_cls_label when label is NULL
- Fix potential buffer overflow in prompt_len underflow (negative-to-size_t wraparound)
- Fix uint16_t type mismatch in audio PCM functions: miniaudio writes unsigned int
  (4 bytes) for channels, but uint16_t (2 bytes) was used, causing stack corruption
- Fix whisper params string lifetime: track default vs dynamically allocated strings
  to avoid sqlite3_free on static string literals
- Fix PCM buffer free: use sqlite3_free instead of free for buffers allocated via
  miniaudio's custom sqlite3_malloc allocator
- Remove redundant NULL assignment in buffer_destroy

Submodule updates:
- Update miniaudio from 0.11.22 to 0.11.25
- Update llama.cpp and whisper.cpp submodules

CI (main.yml):
- Add whisper model and audio test file download with caching
- Rename download-model job to download-models
- Cache whisper model (ggml-tiny.bin) and audio test WAV (jfk.wav) across runs

Makefile:
- Add whisper model and audio test file download targets
- Pass --whisper-model and --audio arguments to unit test binary

Documentation:
- API.md: add llm_chat_system_prompt, audio_model_load, audio_model_free,
  audio_model_transcribe with full options table; fix version and typos
- README.md: move Getting Started after Features; add examples for embedding
  generation, chat, and audio transcription; add Audio Transcription feature;
  update description to reflect shipped audio support

Version bump: 0.7.59 → 0.9.0

+++ src/utils.h
@@ -48,11 +48,11 @@ void buffer_destroy (buffer_t *b);
 
 char *ai_uuid_v7_string (char value[UUID_STR_MAXLEN], bool dash_format);
 
-float *audio_wav_file2pcm (const char *wav_path, uint64_t *num_samples, uint32_t *sample_rate, uint16_t *channels);
-float *audio_wav_mem2pcm (const void *data, size_t data_size, uint64_t *num_samples, uint32_t *sample_rate, uint16_t *channels);
-float *audio_flac_file2pcm (const char *flac_path, uint64_t *num_samples, uint32_t *sample_rate, uint16_t *channels);
-float *audio_flac_mem2pcm (const void *data, size_t data_size, uint64_t *num_samples, uint32_t *sample_rate, uint16_t *channels);
-float *audio_mp3_file2pcm (const char *mp3_path, uint64_t *num_samples, uint32_t *sample_rate, uint16_t *channels);
+float *audio_wav_file2pcm (const char *wav_path, uint64_t *num_samples, uint32_t *sample_rate, uint32_t *channels);
+float *audio_wav_mem2pcm (const void *data, size_t data_size, uint64_t *num_samples, uint32_t *sample_rate, uint32_t *channels);
+float *audio_flac_file2pcm (const char *flac_path, uint64_t *num_samples, uint32_t *sample_rate, uint32_t *channels);
+float *audio_flac_mem2pcm (const void *data, size_t data_size, uint64_t *num_samples, uint32_t *sample_rate, uint32_t *channels);
+float *audio_mp3_file2pcm (const char *mp3_path, uint64_t *num_samples, uint32_t *sample_rate, uint32_t *channels);
 float *audio_mp3_mem2pcm (const void *data, size_t data_size, uint64_t *num_samples, uint32_t *sample_rate, uint32_t *channels);
 int    audio_list_devices (void *xdata, audio_list_devices_callback input_devices_cb, audio_list_devices_callback output_devices_cb);
